Tweak API to couple date, day, and humanized date to TaskList
Since we are trying to create an effortless API that makes the TaskList the
primary view-model in the application, this is an important step. What this
coupling does is that it allows us to update the entire UI with a single change
in the view-model. It will change the date-markers and the humanized versions
of the dates as well.